+++ CHANGELOG.md
@@ -9,7 +9,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
-- Minimal implementation of bucket and object resources.
+- Minimal implementation of bucket, bucket policy and object resources.
 
 [Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/releases/tag/v0.0.1